chore: expose session id setter
